% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cwf_cache_enable}
\alias{cwf_cache_enable}
\title{Enable on-disk caching of CHILDES queries}
\usage{
cwf_cache_enable(cache_dir = NULL)
}
\arguments{
\item{cache_dir}{Directory for cached results; defaults to user cache dir.}
}
\description{
Enable on-disk caching of CHILDES queries
}
